amended home page, amended navbar, tested background images, need to change service data model

+++ Clean-server/app/__init__.py
@@ -57,46 +57,57 @@ def readRooms():
 
 @app.route("/handleJobCreation/<string:jobHash>", methods=["GET", "POST"])
 def handleJobCreation(jobHash):
+    stagedJobSession = models.ST01_HandleJobCreation.objects(D1002=jobHash)
+    stagedJob = stagedJobSession.first()
 
     if request.method == "POST":
-        data = request.json
+        if not stagedJob:
+            data = request.json
 
-        hashSet = "J9z0#HZf5a&z$J@T4d1s29dc5s3J$0TYf@2z&*7bx28g3Z1d!#01vkur7jz^J4Erv61Y7$Y5Q!SO2w6e6W"
-        recId = "".join(random.sample(hashSet, 10))
+            hashSet = (
+                "J9z0HZf5azJT4d1s29dc5s3J0TYf2z7bx28g3Z1d01vkur7jzJ4Erv61Y7Y5QSO2w6e6W"
+            )
+            recId = "".join(random.sample(hashSet, 10))
 
-        new_record = models.ST01_HandleJobCreation(**data)
-        new_record.D1002 = recId
+            new_record = models.ST01_HandleJobCreation(**data)
+            new_record.D1002 = recId
+            new_record.save()
 
-        new_record.save()
+            coll = models.ST01_HandleJobCreation.objects
 
-        coll = models.ST01_HandleJobCreation.objects
+            updatedRec = coll(D1002=recId)
+            updatedRec = updatedRec.first()
 
-        updatedRec = coll(D1002=recId)
-        updatedRec = updatedRec.first()
+            return jsonify({"response": updatedRec.D1002})
+        else:
+            stagedJob.delete()
+            new_record = models.ST01_HandleJobCreation(**data)
+            new_record.save()
 
-        return jsonify({"response": updatedRec.D1002})
+            coll = models.ST01_HandleJobCreation.objects
+
+            updatedRec = coll(D1002=jobHash)
+            updatedRec = updatedRec.first()
+
+            return jsonify({"response": updatedRec.D1002})
 
     else:
         stagedJobSession = models.ST01_HandleJobCreation.objects(D1002=jobHash)
-
         stagedJob = stagedJobSession.first()
-        stagedJob.delete()
-
         print(stagedJob)
-
         return jsonify(stagedJob)
 
 
-@app.route("/getST01Temp/<string:jobHash>", methods=["GET"])
-def readSTJobCreationTemp(jobHash):
-    stagingJobCreationSession = models.ST01_HandleJobCreation.objects(D1002=jobHash)
+@app.route("/refreshST01", methods=["POST"])
+def findCustomer():
+    customerID = request.json
+    customerJobs = models.ST01_HandleJobCreation.objects(D1003=customerID)
 
-    tempStagingJobCreation = stagingJobCreationSession.first()
-    tempStagingJobCreation.delete()
-
-    print(tempStagingJobCreation)
-
-    return jsonify(tempStagingJobCreation)
+    if not customerJobs:
+        return jsonify({"error": "no data"})
+    else:
+        customerJobs.delete()
+        return jsonify({"success": f"deleted {customerID} jobs"})
 
 
 @app.route("/getDataTemp/<string:collection>", methods=["GET"])
